`Added Azure deployment files to CRUD application 2`
` `

`Updated the `CRUD application 2.csproj` file to include `deploy.json` and `deploy.parameters.json`. The `deploy.json` file is an Azure Resource Manager (ARM) template that defines the infrastructure and configuration for the project, setting up a server farm and a website on Azure. The `deploy.parameters.json` file provides parameter values for the ARM template, specifying values for the resources to be deployed on Azure.`